+++ sema4ai/src/sema4ai_code/refresh_agent_spec_helper.py
@@ -5,9 +5,6 @@ from typing import TypedDict
 if typing.TYPE_CHECKING:
     from sema4ai_code.agents.list_actions_from_agent import ActionPackageInFilesystem
 
-SEMA4AI = "sema4ai"
-MYACTIONS = "myactions"
-
 
 class ActionPackage(TypedDict, total=False):
     name: str | None
@@ -49,9 +46,7 @@ def _update_agent_spec_with_actions(
             "whitelist": "",
         }
         for action_package in action_packages_in_filesystem
-        if action_package.organization.replace(".", "").lower() in [SEMA4AI, MYACTIONS]
     ]
-
     missing = []
 
     # First try to match by path.