+++ docs/changelog.md
@@ -1,7 +1,8 @@
 ## Unreleased
 
+- Add `sema4ai.addMCPServer` command
 - Disable notifications for computing datasource status in background.
-- Update Agent CLI to `v1.3.4`
+- Update Agent CLI to `v2.0.2`
 - Enabled new organizations to be part of the Agent spec & the Actions folder in the Agent Package
 - Support for the `v3` of the Agent spec.
   - Adds support for defining `mcp-servers` in the `agent-spec.yaml`.